Test various FTP directory listing formats and fix omission
I have added a test which calls the UnixDirectoryParser with a collection of FTP directory listing which have been acquired from many differing sources. As always IIS is the one which breaks with FTP tradition and doesn't follow the normal time format. I have added two more date/time formats in ExtractFtpDate which cater for single digit hours which are unique to the IIS FTP server.