Suport Dependency Injection Frameworks
Create an interface definition IFtpClient for FtpClient and derive FtpClient from it.

Add a parameterless constructor and a Configure() method so that a dependency injection framework can make an IFtpClient and the consumer of that can configure the injected class. The original constructor is still available for use in non-DI environments.